# java-spring/src/main/java/org/openapitools/api/IlluminaApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.RecipeIdentifier;
import org.openapitools.model.SequencingRunMetrics;
import org.openapitools.model.SequencingRunStatusDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-08-08T00:23:41.052369500Z[Etc/UTC]")
@Validated
@Tag(name = "Lims", description = "the Lims API")
public interface IlluminaApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /Illumina/Sequencer/v2/sequencing-run/login
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "limsGetLoginUrl",
        tags = { "Lims" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/xml", schema = @Schema(implementation = Object.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/Illumina/Sequencer/v2/sequencing-run/login",
        produces = { "application/json", "text/json", "application/xml", "text/xml" }
    )
    default ResponseEntity<Object> limsGetLoginUrl(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /Illumina/Sequencer/v2/sequencing-run/recipe/novaseq : Get the recipe
     *
     * @param recipeIdentifier  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "limsGetRecipe",
        summary = "Get the recipe",
        tags = { "Lims" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/xml", schema = @Schema(implementation = Object.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/Illumina/Sequencer/v2/sequencing-run/recipe/novaseq",
        produces = { "application/json", "text/json", "application/xml", "text/xml" },
        consumes = { "application/json", "text/json", "application/xml", "text/xml", "application/x-www-form-urlencoded" }
    )
    default ResponseEntity<Object> limsGetRecipe(
        @Parameter(name = "RecipeIdentifier", description = "", required = true) @Valid @RequestBody RecipeIdentifier recipeIdentifier
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /Illumina/Sequencer/v2/sequencing-run/recipe/samplesheetfile
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "limsGetSampleSheet",
        tags = { "Lims" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/xml", schema = @Schema(implementation = Object.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/Illumina/Sequencer/v2/sequencing-run/recipe/samplesheetfile",
        produces = { "application/json", "text/json", "application/xml", "text/xml" }
    )
    default ResponseEntity<Object> limsGetSampleSheet(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /Illumina/Sequencer/v2/sequencing-run/metrics
     * Submits sequencing run metrics.
     *
     * @param sequencingRunMetrics  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "limsSubmitMetrics",
        description = "Submits sequencing run metrics.",
        tags = { "Lims" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/xml", schema = @Schema(implementation = Object.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/Illumina/Sequencer/v2/sequencing-run/metrics",
        produces = { "application/json", "text/json", "application/xml", "text/xml" },
        consumes = { "application/json", "text/json", "application/xml", "text/xml", "application/x-www-form-urlencoded" }
    )
    default ResponseEntity<Object> limsSubmitMetrics(
        @Parameter(name = "SequencingRunMetrics", description = "", required = true) @Valid @RequestBody SequencingRunMetrics sequencingRunMetrics
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /Illumina/Sequencer/v2/sequencing-run/update
     * Updates run progress.
     *
     * @param sequencingRunStatusDto  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "limsUpdateProgress",
        description = "Updates run progress.",
        tags = { "Lims" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/json", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Object.class)),
                @Content(mediaType = "text/xml", schema = @Schema(implementation = Object.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/Illumina/Sequencer/v2/sequencing-run/update",
        produces = { "application/json", "text/json", "application/xml", "text/xml" },
        consumes = { "application/json", "text/json", "application/xml", "text/xml", "application/x-www-form-urlencoded" }
    )
    default ResponseEntity<Object> limsUpdateProgress(
        @Parameter(name = "SequencingRunStatusDto", description = "", required = true) @Valid @RequestBody SequencingRunStatusDto sequencingRunStatusDto
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}